Slop manual e2e setup
